[GOB-21] - convert all parseInts to parseFloats and removed unused import

+++ backend/src/utils/restaurant.util.ts
@@ -1,16 +1,16 @@
-import { GetRestaurantByFiltersResponse, MenuItem, Restaurant } from "../types/restaurant.type";
+import { GetRestaurantByFiltersResponse, MenuItem } from "../types/restaurant.type";
 import { get } from "lodash";
 
 export const groupMenuItemDataByRestaurant = (menuItemData: GetRestaurantByFiltersResponse[]) => {
     return Object.values(menuItemData.reduce((groupedMenuItemData: any, restaurant: GetRestaurantByFiltersResponse) => {
 
         const menuItems: MenuItem[] = get(groupedMenuItemData, `${restaurant.restaurantName}.menuItems`, []);
-        menuItems.push({ name: restaurant.menuItemName, price: parseInt(restaurant.menuItemPrice) });
+        menuItems.push({ name: restaurant.menuItemName, price: parseFloat(restaurant.menuItemPrice) });
 
         groupedMenuItemData[restaurant.restaurantName] = {
             name: restaurant.restaurantName,
-            lat: parseInt(restaurant.latitude),
-            long: parseInt(restaurant.longitude),
+            lat: parseFloat(restaurant.latitude),
+            long: parseFloat(restaurant.longitude),
             menuItems
         };
